requirements.txt
gspread==5.4.0
Django==4.0.5
djangorestframework==3.12.4
psycopg2-binary==2.9.3
pytz==2022.1
google-auth-oauthlib==0.5.2
google-api-python-client==2.51.0
beautifulsoup4==4.11.1
python-dotenv==0.20.0
requests==2.28.0